requirements.txt
selenium==3.141.0
mysql-connector==2.2.9
